Extract generic Observable functions into a separate file
Extract functions for generating and manipulating observable
streams which are not text-selection-specific out of `selections.js`
into a separate utility module.
